+++ packages/common/src/client/sync/stream/AbstractStreamingSyncImplementation.ts
@@ -899,6 +899,9 @@ The next upload iteration will be delayed.`);
     let receivingLines: Promise<void> | null = null;
     let hadSyncLine = false;
 
+    if (signal.aborted) {
+      throw new AbortOperation('Connection request has been aborted');
+    }
     const abortController = new AbortController();
     signal.addEventListener('abort', () => abortController.abort());
 
@@ -1071,7 +1074,9 @@ The next upload iteration will be delayed.`);
       await control(PowerSyncControlCommand.START, JSON.stringify(options));
 
       this.notifyCompletedUploads = () => {
-        controlInvocations?.enqueueData({ command: PowerSyncControlCommand.NOTIFY_CRUD_UPLOAD_COMPLETED });
+        if (controlInvocations && !controlInvocations?.closed) {
+          controlInvocations.enqueueData({ command: PowerSyncControlCommand.NOTIFY_CRUD_UPLOAD_COMPLETED });
+        }
       };
       await receivingLines;
     } finally {