+++ packages/adapter-sql-js/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @powersync/adapter-sql-js
 
+## 0.0.4
+
+### Patch Changes
+
+- 47294f2: Update PowerSync core extension to version 0.4.4
+- Updated dependencies [876c550]
+- Updated dependencies [c910c66]
+- Updated dependencies [9e3e3a5]
+  - @powersync/common@1.37.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [7609155]
+- Updated dependencies [7ad251a]
+- Updated dependencies [7f2c53d]
+  - @powersync/common@1.36.0
+
+## 0.0.2
+
+### Patch Changes
+
+- 8cc3335: Updated sql-js upstream version to support rust sync implementation.
+- 074f4cd: Only calling `db.export()` if a persister is specified, otherwise it is no-op. Improves in-memory performance.
+
 ## 0.0.1
 
 ### Patch Changes