+++ packages/react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @powersync/react
 
+## 1.7.1
+
+### Patch Changes
+
+- 8decd49: Refactor useQuery hook to avoid calling internal hooks conditionally.
+- Updated dependencies [876c550]
+- Updated dependencies [c910c66]
+- Updated dependencies [9e3e3a5]
+  - @powersync/common@1.37.0
+
+## 1.7.0
+
+### Minor Changes
+
+- 7ad251a: Updated package exports to reflect ESM nature.
+- 7ad251a: Update package exports to reflect ESM exports.
+
+### Patch Changes
+
+- dce523a: Fixed regression in useSuspendingQuery where `releaseHold is not a function` could be thrown during rendering.
+- Updated dependencies [7609155]
+- Updated dependencies [7ad251a]
+- Updated dependencies [7f2c53d]
+  - @powersync/common@1.36.0
+
 ## 1.6.0
 
 ### Minor Changes