+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @powersync/vue
 
+## 0.4.0
+
+### Minor Changes
+
+- 7ad251a: Updated package exports to reflect ESM nature.
+
+### Patch Changes
+
+- Updated dependencies [7609155]
+- Updated dependencies [7ad251a]
+- Updated dependencies [7f2c53d]
+  - @powersync/common@1.36.0
+
 ## 0.3.0
 
 ### Minor Changes